Refactor NewsArticleAuthor
Uses flexbox for layout; allows the component to be responsive without
media queries.

Replace Author component with AuthorName and ItalicText components.
These components are contained within Author, but they needed some
extra styling to make them work with the new layout, so they were
broken out. It’s possible that the Author component can be removed in
favor of this approach.