Make mobileTitle inline-block and add top padding
inline-block is used instead of block so that the title only
takes up as much width as it needs based on the amount of
text. Otherwise, the title spans the entire width of the
available space and it doesn't really need to.

Top padding is used to push the text down and center it
vertically.