Add new values for z-index
There needs to be more variation and intention with the z-index values.
Currently, many components use “modal” and then add or subtract values
to place the component on the desired layer. The addition of the global
log in modal has revealed that this method for setting z-index is not
very sustainable. Essentially, “modal” doesn’t mean much at this point
because too many components that aren’t a modal use it.

The goal here is to introduce some hierarchy within the z-index values.
Every component that sits on top of the page should not use “modal” or
“modal + 1”, because when an actual modal is rendered now, it may or
may not be placed on top of every other component. The stacking order
of components should be clearly defined and this should help.