Set input opacity to 0 to hide it
In legacy versions of IE, the checkbox input type cannot be
styled and the appearance property is not supported. Because
of this, there was a bug in IE where the fake input’s border
and the actual input’s border both appeared.

Setting the input’s opacity to 0 fixes this bug by hiding
the element from view, but still allowing it to be
interacted with. The input can safely be hidden because the
fake input gives the visual cue as to where the checkbox is
and what it looks like.