Refactor Typeahead component
Previously, the Typeahead component used a forked version of
react-typeahead which was installed via a Git repo URL. This
was not ideal and caused issues when deploying Docker apps.

The solution here is to take the source code from the forked
version of react-typeahead and include it here as the
Typeahead component.