Scroll to scrollPosition state when the modal open state is false
When the modal is closed, the window is scrolled back to where the
user was before the modal was opened. This creates a better experience
and prevents an annoying scroll jump when the modal is closed. This is
just the value that is calculated by noScroll.on() and stored in the
state when the modal opens.