Change role from alert to alertdialog in Toast component
> If an alert also provides interactive controls (such as form controls
that allow the user to rectify a problem, or an "OK" button that
discards the alert) the alertdialog role should be used instead.

Based on the spec above, the Toast component would be considered an
alertdialog because of the close/dismiss button that is present. Also,
in some Toast messages, there are links; for example, "View your lists"
after adding a new bookmark.