feat: Improve accessibility of checkbox component
There was a significant bug where the checkbox element could not be
tabbed to or selected with the keyboard.

A native HTML checkbox can be toggled when focused with the spacebar
key. This update adds that functionality back in.

This component itself may need to be re-thought and re-built from
an accessibility perspective, for not only keyboard users, but also
for screen readers.