refactor: DateRange component outside range test
This began as a bug fix but I think I determined that it was in
actuality an issue with the DateRangePicker dependency and not our
code, after implementing tests for our code. Adding tests allowed me to
refactor in a way that seems a bit more straightforward. Also fixed an
issue with the component that was firing React error messages in the
console.

Introduced enzyme as a library to help test React components, and
refactored existing tests a bit for improved visualization.