fix: console errors, updated validate dep
Fixed several console issues related to using the input and Validate
component, by cleaning up which props could be spread on the input.
Unfortunately this was spreading dom event handlers like onBlur also,
and that's expected behavior even though not defined in the Input api.

Additionally, updated the version of the react-validate-form so we no
longer get dependency errors for that package.