chore: fork react-motion-ui-pack (#741)
* chore: fork react-motion-ui-pack

This library depended on a small package that was no longer maintained.
I've created a libs folder and moved over the relevant pieces of this
library that we use. In the future, we should migrate to the use of a
supported package or deprecate this functionality altogether, however I
thought this would be the best short term way to remove the npm warnings
while still maintaining backwards compatibility, and also allowing us to
make small changes in the meantime. Willing to change direction on this
if necessary.

* chore: remove get-prefix dep by same maintainer